Refactor: move definitions from `zones` to `zones::types`
